Fix fr chapter 4 (#513)
* Conjugation fix

Fixed an error when "continuer" had a mistake.

* Fix a grammar mistake

* Fixed minor grammar mistake

* Fixed minor grammar and expression issues

* Made sentence easier to read.

* fixed typo

* Fixed minor typo

* Fixed typos and grammar mistakes

* fixed minor grammar issues and typos

I deleted the translation of "for" loop because it doesn't help tos. understand the concept of this kind of loop. Plus, this is a very basic functionality in almost all development languages.
I also replaced the word "extérieurement" by the expression "depuis l'extérieur". It is more common to say it like that in French.

* changed some sentences for better understanding

* fixed minor grammar issue

* Minor grammar fixes

* Fixed typo

* Fixed minor typo